Refactored time registration entry.
Now the timestamp is pieced together from hh:mm and a date Y-m-d - this corresponds to the time registration form.
Work is now needed for the part once the registration has been done, and showing the entries already registered on the day - as well as date navigation to go back and forth.