More work on the time registration form
The form is coming along,
The activitiy list is now fixed, and shows correct activities for a user.
Needs work on having the select elements selecting the activitiy chosen for that row. Once that is figured out the start and end time for the activitiy needs to be pre-selected as well.